Copy string to new memory space

+++ 0x0B-malloc_free/1-strdup.c
@@ -11,7 +11,7 @@
 char *_strdup(char *str)
 {
 	char *copyString = malloc(sizeof(char) * strlen(str));
-	long unsigned int i = 0;
+	unsigned int i = 0;
 
 	if (copyString == NULL || str == NULL)
 	{
